Add auto north-up mode, disable idle timer, and heading smoothing
- Add device pitch detection using CoreMotion
- Auto-switch to north-up when device tilted > 70 degrees (for car mounts)
- Disable idle timer while in map view to prevent screen lock
- Add heading smoothing with low-pass filter for stable compass
- Rate limit heading updates and filter low-accuracy readings
- Archive old project structure

+++ sweTAK iOS/sweTAK iOS/ViewModels/RoutesViewModel.swift
@@ -143,12 +143,12 @@ public final class RoutesViewModel: ObservableObject {
     }
 
     /// Total distance of all breadcrumb routes in meters
-    public var totalBreadcrumbDistance: Float {
+    public var totalBreadcrumbDistance: Double {
         breadcrumbRoutes.reduce(0) { $0 + $1.totalDistanceMeters }
     }
 
     /// Total distance of all planned routes in meters
-    public var totalPlannedDistance: Float {
+    public var totalPlannedDistance: Double {
         plannedRoutes.reduce(0) { sum, route in
             sum + route.totalDistanceMeters
         }
